fix _editUiState 관련 에러 해결

+++ app/src/main/java/com/example/dodum_android/network/contest/ContestService.kt
@@ -28,11 +28,11 @@ interface ContestService {
 
     @DELETE(DodumUrl.Contest.ID)
     suspend fun deleteContest(
-        @Path("id") id: Int
+        @Path("id") id: Long
     ): Response<ContestCommonResponse>
 
     @POST(DodumUrl.Contest.ACTIVE)
     suspend fun toggleContestAlert(
-        @Path("id") id: Int
+        @Path("id") id: Long
     ): Response<ContestActiveResponse>
 }
